+++ NetService/src/SQLUserStorage.h
@@ -9,7 +9,7 @@ public:
 	sUserStorage ();
 	~sUserStorage ();
 	bool IfUserExist(const char* pUser);
-	bool GetPassword(const char* pUser, char* pPassWordBuff, long cbBuff);
+	bool GetPassword(const char* pUser, char* pPassWordBuff, size_t cbBuff);
 	bool GetUserInfo(const char* pUser, userinfo* pUserInfo);
 private:
 	void SaveConnLog(string host,string name,string password,string db,int port /*= 3306*/ );
